Сделан рефракторинг, некоторые проверки, не содержащие запросы к БД, перенесены в сервисы
