+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service;
 
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.dto.FilmDto;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.mapper.FilmMapper;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
@@ -10,6 +11,7 @@ import ru.yandex.practicum.filmorate.storage.FilmRepository;
 import ru.yandex.practicum.filmorate.storage.GenreRepository;
 import ru.yandex.practicum.filmorate.storage.LikeRepository;
 
+import java.time.LocalDate;
 import java.util.*;
 
 @Service
@@ -47,6 +49,7 @@ public class FilmService {
     }
 
     public FilmDto create(Film film) {
+        checkReleaseDate(film);
         Film createdFilm = filmRepository.create(film);
         Map<Long, Set<Genre>> genresByFilm = genreRepository.findGenresForFilms(List.of(createdFilm.getId()));
         createdFilm.setGenres(genresByFilm.getOrDefault(createdFilm.getId(), new LinkedHashSet<>()));
@@ -54,6 +57,7 @@ public class FilmService {
     }
 
     public FilmDto update(Film film) {
+        checkReleaseDate(film);
         Film updatedFilm = filmRepository.update(film);
         Map<Long, Set<Genre>> genresByFilm = genreRepository.findGenresForFilms(List.of(updatedFilm.getId()));
         updatedFilm.setGenres(genresByFilm.getOrDefault(updatedFilm.getId(), new LinkedHashSet<>()));
@@ -84,6 +88,11 @@ public class FilmService {
                 .map(FilmMapper::mapToFilmDto)
                 .toList();
     }
+
+    private void checkReleaseDate(Film film) {
+        if (film.getReleaseDate().isBefore(LocalDate.of(1895, 12, 28))) {
+            throw new ValidationException("Дата выхода не может быть раньше " +
+                    "28.12.1895 - даты выхода первого в истории фильма");
+        }
+    }
 }
-
-
